// healthcare-life-sciences/automate-discharge-instructions/src/types/bedrock-action-group-input.ts
export interface AmazonBedrockActionGroupInput {
  messageVersion: '1.0';
  agent: {
    name: string;
    id: string;
    alias: string;
    version: string;
  };
  inputText: string;
  sessionId: string;
  actionGroup: string;
  apiPath: string;
  httpMethod: string;
  parameters: {
    name: string;
    type: string;
    value: string;
  }[];
  requestBody: {
    content: {
      [contentType: string]: {
        properties: {
          name: string;
          type: string;
          value: string;
        }[];
      };
    };
  };
  sessionAttributes: {
    [key: string]: string;
  };
  promptSessionAttributes: {
    [key: string]: string;
  };
}